agregando la funcionalidad de kg,gramos y unidad y el diseño

+++ src/main/java/com/baez/baezpos/inventory/dto/InventoryMovementRequest.java
@@ -3,10 +3,12 @@ package com.baez.baezpos.inventory.dto;
 import com.baez.baezpos.inventory.entity.MovementType;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class InventoryMovementRequest {
     private Long productId;
-    private Integer quantity;
+    private BigDecimal quantity; // BigDecimal para soportar movimientos fraccionados
     private MovementType type;
     private String reason;
 }
