+++ src/main/java/com/baez/baezpos/inventory/dto/InventoryMovementResponseDTO.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -17,7 +18,7 @@ public class InventoryMovementResponseDTO {
     private Long productId;
     private String productName;
     private MovementType movementType;
-    private Integer quantity;
+    private BigDecimal quantity; // BigDecimal para mostrar fracciones en historial de inventario
     private String reason;
     private LocalDateTime createdAt;
 }
