+++ src/main/java/com/baez/baezpos/sale/dto/SaleItemRequestDTO.java
@@ -4,6 +4,6 @@ import java.math.BigDecimal;
 
 public record SaleItemRequestDTO(
         Long productId,
-        Integer quantity,
+        BigDecimal quantity, // BigDecimal para soportar fracciones: 0.250 kg, 1.500 kg, etc.
         BigDecimal price
 ) {}
